requirements.txt
streamlit
langchain
langchain_community

python_dotenv
duckduckgo-search
arxiv
wikipedia
google-search